6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /**
  * main - entry point
- * Return 0 (success)
+ * Return: 0 (success)
  */
 int main(void)
 {
@@ -10,10 +10,10 @@ int b;
 long d;
 long long e;
 float g;
-printf("Size of a char:%lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int:%lu byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int:%lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a long long byte:%lu byte(s)\n", (unsigned long)sizeof(e));
-printf("Size of a float:%lu byte(s)\n", (unsigned long)sizeof(g));
+printf("Size of a char:%zu byte(s)\n", sizeof(a));
+printf("Size of an int:%zu byte(s)\n", sizeof(b));
+printf("Size of a long int:%zu byte(s)\n", sizeof(d));
+printf("Size of a long long byte:%zu byte(s)\n", sizeof(e));
+printf("Size of a float:%zu byte(s)\n", sizeof(g));
 return (0);
 }
